+++ c++/utility.hpp
@@ -6,7 +6,7 @@
 typedef double price_type;
 typedef double size_type;
 typedef int id_type;
-typedef unsigned time_type;
+typedef long time_type;
 
 extern id_type ID_DEFAULT;
 extern size_type SIZE_DEFAULT;
